requirements.txt
amqp==5.3.1
asgiref==3.8.1
azure-core==1.34.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
django==5.1.5
django-environ==0.12.0

django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.1
mysqlclient==2.2.7
mysql-connector-python==9.2.0
numpy==1.26.3

pandas==2.2.3
packaging==24.2
prompt-toolkit==3.0.50

pymysql==1.1.1
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
whitenoise==6.9.0
